SIT adding themself as mentors with same email bug fix

+++ app/services/mentors/transfer_existing_mentor_to_sit.rb
@@ -7,11 +7,13 @@ module Mentors
     def call
       ActiveRecord::Base.transaction do
         mentor_user = mentor_profile.user
-        Identity::Transfer.call(from_user: sit_user, to_user: mentor_user)
         Induction::Enrol.call(participant_profile: mentor_profile, induction_programme: school_cohort.default_induction_programme, start_date:)
         Mentors::AddToSchool.call(mentor_profile:, school: school_cohort.school)
-        sit_user.reload.destroy!
-        mentor_profile.reload
+        if mentor_user != sit_user
+          Identity::Transfer.call(from_user: sit_user, to_user: mentor_user)
+          sit_user.reload.destroy!
+          mentor_profile.reload
+        end
       end
     end
